#7 refresh stream implemented (#16)
* added skeleton button, rename

renamed  close button and refresh to differentiate, added empty function with console output. Still determining if closing stream and reopening or investigate if there is internal refresh

* complete refresh implementation

refresh will delete and recreate the embedded player. This is triggered by both retyping the stream name into the search box or manually pressing the "re".

This will remove the embedded player and remake it. Buttons are left alone, looks like they are shifted over to the new location without need of remaking them as well.